[Refactor] encapsulate the drop cursor rendering logic into `dropCursorRenderer.ts`
